Initial move of the yoast-components-app
