+++ apps/components/ComponentsExample.js
@@ -2,11 +2,11 @@ import React from "react";
 import styled from "styled-components";
 import { __ } from "@wordpress/i18n";
 
-import YoastWarning from "../composites/Plugin/Shared/components/YoastWarning";
-import { FullHeightCard } from "../composites/CoursesOverview/Card";
-import CardDetails from "../composites/CoursesOverview/CardDetails";
-import getCourseFeed from "../utils/getCourseFeed";
-import { getRtlStyle } from "../utils/helpers/styled-components";
+import YoastWarning from "../../packages/yoast-components/composites/Plugin/Shared/components/YoastWarning";
+import { FullHeightCard } from "../../packages/yoast-components/composites/CoursesOverview/Card";
+import CardDetails from "../../packages/yoast-components/composites/CoursesOverview/CardDetails";
+import getCourseFeed from "../../packages/yoast-components/utils/getCourseFeed";
+import { getRtlStyle } from "../../packages/yoast-components/utils/helpers/styled-components";
 
 const Container = styled.div`
 	max-width: 1024px;
