+++ apps/components/HelpCenterWrapper.js
@@ -1,10 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 
-import HelpCenter from "../composites/Plugin/HelpCenter/HelpCenter";
-import VideoTutorial from "../composites/HelpCenter/views/VideoTutorial";
-import AlgoliaSearcher from "../composites/AlgoliaSearch/AlgoliaSearcher";
-import colors from "../style-guide/colors.json";
+import HelpCenter from "../../packages/yoast-components/composites/Plugin/HelpCenter/HelpCenter";
+import VideoTutorial from "../../packages/yoast-components/composites/HelpCenter/views/VideoTutorial";
+import AlgoliaSearcher from "../../packages/yoast-components/composites/AlgoliaSearch/AlgoliaSearcher";
+import colors from "../../packages/yoast-components/style-guide/colors.json";
 
 export const HelpCenterContainer = styled.div`
 	max-width: 1280px;
