+++ apps/components/reducers/index.js
@@ -1,0 +1,7 @@
+import { combineReducers } from "redux";
+
+import contentAnalysis from "../../../packages/yoast-components/composites/Plugin/ContentAnalysis/reducers/contentAnalysisReducer";
+
+export default combineReducers( {
+	contentAnalysis,
+} );
