--- packages/yoast-components/app/reducers/index.js
@@ -1,7 +1,0 @@
-import { combineReducers } from "redux";
-
-import contentAnalysis from "../../composites/Plugin/ContentAnalysis/reducers/contentAnalysisReducer";
-
-export default combineReducers( {
-	contentAnalysis,
-} );
